REFACTOR: zustand persist #17 (추가: 스토어 네이밍 변경)
