+++ src/zustand/searchStore.ts
